+++ NEWS.md
@@ -4,7 +4,7 @@ This page describes the most important changes in `Backport`. The format is base
 a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
 Versioning](https://semver.org).
 
-## Unreleased
+## Version 0.1.6 (2025-05-28)
 
 ### Changed